vmodtool: Add option to specify c names of arguments and avoid "bool"
I tried gcc version 15.0.0 20241203 and even without -std=c23, it would reserve
"bool":

In file included from vcc_std_if.c:10:
vcc_std_if.h:78:33: error: two or more data types in declaration specifiers
   78 |         VCL_BOOL                bool;
      |                                 ^~~~

(and more)

So this patch adds the option to specify the c name of arguments to vmod
functions/methods by separating it with a colon in the spec in the VCC file.

	vclname:cname

We use this facility to rename the "bool" argument to vmod_std functions to
"boola".

Implementation
--------------

The cname needs to be specified by the VMOD author because it is also used in
the vmod implementation.

Obviously, VCC needs to know that name, too, to generate argument structs, so we
need to transport it from the VMOD to VCC via the JSON spec. A logical place for
cname is next to the name, because the other attributes following in the
argument spec array are optional. So this changes the JSON spec format, and,
consequently, we need to increase the version number.

So, ultimately, this is a breaking change with respect to vmodtool: One can not
import vmods built before this change. We could add compatibility for this case,
but as we have a tradition of forcing rebuilds with each release by bumping the
VRT major number anyway, I did not see my time well spent on implementing
backwards compatibility.

Fixes #4294